Convert memory and disk on the ProcessUsage object to Long
This is a follow up to PR #1085 and was pointed out in PR #1064.

+++ cloudfoundry-client/src/main/java/org/cloudfoundry/client/v3/processes/_ProcessUsage.java
@@ -37,13 +37,13 @@ abstract class _ProcessUsage {
      * The disk
      */
     @JsonProperty("disk")
-    abstract Integer getDisk();
+    abstract Long getDisk();
 
     /**
      * The memory
      */
     @JsonProperty("mem")
-    abstract Integer getMemory();
+    abstract Long getMemory();
 
     /**
      * The time
